requirements.txt
PyQt5
ftfy
requests
paho-mqtt
python-dotenv